In the workbench one can define views...
... and this commit adds a class and a parser to handle those views.
In the workbench one has to provide a SQL definition for the view.
As the plain SQL is stored in the workbench file, a new parser class
has to be added to parse that SQL. Appropriate patches are sent to
SQL::Translator and as soon as these changes are merged, the new
parser class should be removed.

Currently the views should avoid table aliases in the SQL definition
and the views shouldn't be too complex. Otherwise the parser has to
be changed to handle those views.